requirements.txt
# Python依赖
paddlepaddle-gpu>=2.0.0
paddlepaddle-gpuvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.14.0
matplotlib>=3.7.0
shapely>=2.0.0  # 用于IoU计算

# 可选：用于更精确的角点检测
scikit-image>=0.21.0

# 开发工具
jupyter>=1.0.0
ipython>=8.14.0
